Converters extension points are factories. Consistent converters names.
