+++ src/main/java/com/github/kornilova_l/flamegraph/plugin/server/converters/calltraces/FileToCallTracesConverterBase.kt
@@ -1,0 +1,8 @@
+package com.github.kornilova_l.flamegraph.plugin.server.converters.calltraces
+
+import java.io.File
+
+/**
+ * @author Liudmila Kornilova
+ **/
+abstract class FileToCallTracesConverterBase(protected val file: File) : FileToCallTracesConverter
