+++ src/main/java/com/github/kornilova_l/flamegraph/plugin/server/converters/cflamegraph/ProfilerToCFlamegraphConverter.kt
@@ -1,0 +1,8 @@
+package com.github.kornilova_l.flamegraph.plugin.server.converters.cflamegraph
+
+/**
+ * @author Liudmila Kornilova
+ **/
+interface ProfilerToCFlamegraphConverter {
+    fun convert(): CFlamegraph
+}
