+++ src/main/java/com/github/kornilova_l/flamegraph/plugin/server/converters/flamegraph/ProfilerToFlamegraphConverter.kt
@@ -1,0 +1,6 @@
+package com.github.kornilova_l.flamegraph.plugin.server.converters.flamegraph
+
+@Deprecated("Implement ProfilerToCFlamegraphConverter instead")
+interface ProfilerToFlamegraphConverter {
+    fun convert(): Map<String, Int>
+}
